parser_project/requirements.txt
httpx>=0.27.0
kafka-python>=2.0.2
pydantic>=2.8.0
python-dotenv>=1.0.1
psycopg[binary]>=3.2.0
feedparser>=6.0.11
PyYAML>=6.0.2
